% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3BarGraphV2.R
\name{d3BarGraphV2-shiny}
\alias{d3BarGraphV2-shiny}
\alias{d3BarGraphV2Output}
\alias{renderD3BarGraphV2}
\title{Shiny bindings for d3BarGraphV2}
\usage{
d3BarGraphV2Output(outputId, width = "auto", height = "400px")

renderD3BarGraphV2(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a d3BarGraphV2}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using d3BarGraphV2 within Shiny
applications and interactive Rmd documents.
}

