% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3BarGraphV2.R
\name{d3BarGraphV2}
\alias{d3BarGraphV2}
\title{d3BarGraph Version 2...putting the plotting function inside of HTMLWidgets.widget}
\usage{
d3BarGraphV2(message, width = NULL, height = NULL)
}
\description{
<Add Description>
}

